RemoveAllSpacesReturnList
This function takes a string and removes the spaces and put the values into a list. If the words are separated by a hyphen or underscore, this formatting is preserved and the long string is added to the list. The problem I was facing was a word with a space before and after the hyphen. Therefore, I parsed out the spaces and concatenated the hyphen with the words before and after it and placed it into the list.