GetSearchData
This function accepts column1 as the column name to narrow the search and column1value as the value (ex. Name (column) "Sarah" (value)), column2 is the target column for the value you're searching for and searchCriteria is the value you're searching for (ex. Address (column) "Main" (searchCriteria)), the tableName is the table where the data resides and connection string holds the database connection. The return value is a DataTable holding the search results.