SaveExcelFile
This function saves DataTable data to an Excel file. The parameters are the full path and filename and the DataTable of the data. The return value is a Boolean of True or False whether the save was successful or not. The supporting functions are the "GetWindowThreadProcessId" function that gets the Excel process, the "ReleaseComObject" releases the process after the Excel file is created and the "AddExcelValues" function fills the worksheet with values from the DataTable.