AddDataToCsvFile
This function adds a DataTable to a .csv file using the parameters of a filename (path and filename) and a DataTable with the values to add to the .csv file. This function uses a Private function "GetCounter" which is used to format the output. The return value is a Boolean datatype that returns True if successful.