GetNextID
This function retrieves all the primary key integer IDs needed from a table based on the demand reflected in the parameter "numRecsNeeded". The parameters are "connectString" (for the connection string of the database), numRecsNeeded (for the number of primary key IDs needed to be returned), pKeyIndex (for the name of the column for the primary key) and tableName (for the name of the target table in the database). If primary keys have been deleted over time (i.e. pKey = 1, 5, 12, 27), then this function will retrieve those deleted IDs and include them with any new generated IDs that are greater than the last know ID and continue to collect them until the full number of IDs requested are added to the list. Then the list is returned.