GetTextFile
This function retrieves a text file and put the values in a datatable and return it. The parameters are "fullPathName" (which is the path and filename of the text file) and "hasHeaders" (a Boolean value determining if headers exist in the file). The "RemoveAllSpacesReturnList" function is necessary to run this function and can be located on this website.