TransferAndOpenFile
This function uses a number of functions and subs to transfer files from one location to another and start the resulting application. The parameters are "localPathAndFile" (which is the full path of the existing application you're targeting) and "existingPathAndFile (which is the name of the application located in a source location). The functions and subs needed for this to work are located on this website which are: 
CopyExistingFilesToLocal, CompareFileVersions,  DeleteLocalFolder, GetCurrentAppLocation and StartApplication.