UpdateXMLElement
This function updates an xml element and saves the file and returns a Boolean value of true if successful. The parameters are "oldValue" (which is the value to be replaced), "newValue" (which is the value that will be added to the element), "eleName" (which is the element where the change will take place), "subRoot" (which is one removed from the root), and "fullFileName" (which is the location of the xml file).