DeleteNode
This sub removes a node from the xml file. The parameters are "xmlFile" (which is the location of the xml file), "subNode" (which is the node removed of the root), "eleNameToDelete" (which is the element where the value will be deleted) and "eleValueToDelete" (which is the value that will be removed).