GetLatestRecords
This function compares the passed in date with the date in the passed in datatable, specified by the dateColumn parameter. If the record in the datatable greater than the passed in date and less than the passed in date plus elapsedTime (which is the number of days you may want to project into the future), then those records are returned.