MakeControlsInvisible
This sub makes all the controls passed in by the parameter "controlContainer" invisible. If only one control is to be made invisible then supply the optional parameter "controlToHide", then only that control will be made invisible. The function GetAllControls is necessary to make this sub work. This function can be found in the Functions section on this website.